+++ src/editor/CanvasController.tsx
@@ -1,17 +1,16 @@
 import { Canvas } from '@react-three/fiber';
 import { OrbitControls } from '@react-three/drei';
 import TorusController from './TorusController';
-import * as THREE from 'three';
 import { JSX } from 'react';
 
 interface CanvasControllerParams {
-  objects: JSX.Element[]
-};
+  objects: JSX.Element[];
+}
 
 function CanvasController({ objects }: CanvasControllerParams) {
   console.log(objects);
   return (
-    <Canvas className='canvas' camera={{ position: [0, 0, 30] }}>
+    <Canvas className="canvas" camera={{ position: [0, 0, 30] }}>
       <ambientLight />
       <directionalLight position={[10, 10, 10]} />
       <TorusController position={[-5, 0, 0]} />
