+++ src/editor/ToolBarItem.tsx
@@ -1,9 +1,15 @@
-import { Children, cloneElement, isValidElement, ReactNode, useState } from "react";
-import { Button, Menu, Fade } from "@mui/material";
+import {
+  Children,
+  cloneElement,
+  isValidElement,
+  ReactNode,
+  useState,
+} from 'react';
+import { Button, Menu, Fade, MenuItem } from '@mui/material';
 
 interface ToolBarItemProps {
-  label: string
-  children?: ReactNode
+  label: string;
+  children?: ReactNode;
 }
 
 function ToolBarItem({ label, children }: ToolBarItemProps) {
@@ -15,22 +21,24 @@ function ToolBarItem({ label, children }: ToolBarItemProps) {
 
   const closeMenu = () => {
     setAnchorEl(null);
-  }
+  };
 
   // modify incoming children, so that each menu item
-  // will also call `closeMenu` on click 
+  // will also call `closeMenu` on click
   const modifiedChildren = Children.map(children, (child) => {
-    if(isValidElement(child)) {
-      const element = child as React.ReactElement<any>;
+    if (isValidElement(child)) {
+      const element = child as React.ReactElement<
+        React.ComponentProps<typeof MenuItem>
+      >;
       const originalOnClick = element.props.onClick;
       return cloneElement(element, {
         onClick: (e: React.MouseEvent<HTMLElement>) => {
           originalOnClick?.(e);
           closeMenu();
-        }
-      })
+        },
+      });
     }
-  })
+  });
 
   return (
     <div className="tool-bar__item">
@@ -52,13 +60,13 @@ function ToolBarItem({ label, children }: ToolBarItemProps) {
         open={open}
         onClose={closeMenu}
         slots={{
-          transition: Fade
+          transition: Fade,
         }}
       >
         {modifiedChildren}
       </Menu>
     </div>
-  )
+  );
 }
 
 export default ToolBarItem;
