+++ src/main.tsx
@@ -2,11 +2,8 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import Home from './pages/Home.tsx'
-import {
-  createBrowserRouter,
-  RouterProvider,
-} from 'react-router-dom';
+import Home from './pages/Home.tsx';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
 const router = createBrowserRouter([
   {
@@ -16,7 +13,7 @@ const router = createBrowserRouter([
   {
     path: '/editor',
     element: <App />,
-  }
+  },
 ]);
 
 createRoot(document.getElementById('root')!).render(
