+++ src/pages/Editor.tsx
@@ -1,21 +1,20 @@
 import { JSX, useState } from 'react';
 import CanvasController from '../editor/CanvasController';
-import Toolbar from '../editor/Toolbar';
-import TorusController from '../editor/TorusController';
+import Toolbar from '../editor/ToolBar';
 
 function Editor() {
   const [objects, setObjects] = useState<JSX.Element[]>([]);
 
   const pushMesh = (mesh: JSX.Element) => {
-    setObjects(prev => [...prev, mesh]);
-  }
+    setObjects((prev) => [...prev, mesh]);
+  };
 
-  return(
+  return (
     <>
       <Toolbar addMesh={pushMesh} />
       <CanvasController objects={objects} />
     </>
-  )
+  );
 }
 
-export default Editor
+export default Editor;
